make code modular
